src/requirements.txt
numpy==1.26.4
pygame-ce==2.4.1
pygame-gui==0.6.9
python-i18n==0.3.9
